Input : add autocomplete

+++ template/manage-member.php
@@ -1,7 +1,7 @@
 <?php 
 
 // Suppression d'un membre
-if($_GET['action'] == 'delete-member' && isset($_GET['id_member'])){
+if(isset($_GET['action']) && $_GET['action'] == 'delete-member' && isset($_GET['id_member'])){
     var_dump($_GET['id_member']);
     $deleteReq = $pdo->prepare("DELETE FROM membre WHERE id_membre = ?");
     $deleteReq->execute(array($_GET['id_member']));
@@ -31,15 +31,15 @@ if(isset($_GET['action']) && ($_GET['action'] == "add-member" || $_GET['action']
     <h2>Gestion des membres</h2>
 
     <form class="col-sm-6" action="inc/function.member.php" method="post">
-        <input type="hidden" class="form-control" name="admin-id-member" id="admin-id-member" placeholder="Votre pseudo" value="<?= $id_member ?>">
+        <input type="hidden" class="form-control" name="admin-id-member" id="admin-id-member" placeholder="Votre pseudo" value="<?= $id_member ?>" autocomplete="family-name">
         <label for="adminpseudo">Pseudo</label><br>
         <input type="text" class="form-control" name="adminpseudo" id="adminpseudo" placeholder="Votre pseudo" value="<?= $pseudo ?>">
         <label for="adminname">Nom</label>
-        <input type="text" class="form-control" name="adminname" id="adminname" placeholder="Votre nom" value="<?= $name?>">
+        <input type="text" class="form-control" name="adminname" id="adminname" placeholder="Votre nom" value="<?= $name?>" autocomplete="family-name">
         <label for="adminfirstname">Prénom</label>
-        <input type="text" class="form-control" name="adminfirstname" id="adminfirstname" placeholder="Votre prénom" value="<?= $firstname ?>">
+        <input type="text" class="form-control" name="adminfirstname" id="adminfirstname" placeholder="Votre prénom" value="<?= $firstname ?>" autocomplete="given-name">
         <label for="adminemail">Email</label>
-        <input type="email" class="form-control" name="adminemail" id="adminemail" placeholder="Votre email" value="<?= $email ?>">
+        <input type="email" class="form-control" name="adminemail" id="adminemail" placeholder="Votre email" value="<?= $email ?>" autocomplete="email">
         <label for="adminsexe">Sexe</label>
         <select class="form-control" name="adminsexe" id="adminsexe">
             <option value="m" <?= ($sex == "m") ? ' selected' : '' ;?> >Homme</option>
